Backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
pillow==10.1.0

google-auth==2.23.4
google-auth-oauthlib==1.1.0

PyJWT==2.8.0